refactor: move load logic to separate module
