+++ iati_tables/__init__.py
@@ -1,4 +1,3 @@
-import concurrent.futures
 import functools
 import json
 import logging
@@ -8,19 +7,15 @@ import time
 from collections import defaultdict
 from datetime import datetime
 from io import StringIO
-from itertools import islice
-from typing import Any, Iterator, Optional, OrderedDict
+from typing import Any, Iterator, Optional
 
-import iatikit
 import requests
-import xmlschema
-from lxml import etree
 from sqlalchemy import column, insert, table, text
 
-from iati_tables import sort_iati
-from iati_tables.database import create_field_sql, get_engine, schema
+from iati_tables.database import create_field_sql, get_engine
 from iati_tables.export import export_all
 from iati_tables.extract import extract
+from iati_tables.load import IATISchemaWalker, load
 from iati_tables.upload import upload_all
 
 logging.basicConfig(
@@ -35,12 +30,6 @@ this_dir = pathlib.Path(__file__).parent.resolve()
 
 output_dir = os.environ.get("IATI_TABLES_OUTPUT", ".")
 output_path = pathlib.Path(output_dir)
-
-
-VERSION_1_TRANSFORMS = {
-    "activity": etree.XSLT(etree.parse(str(this_dir / "iati-activities.xsl"))),
-    "organisation": etree.XSLT(etree.parse(str(this_dir / "iati-organisations.xsl"))),
-}
 
 
 def _create_table(table, con, sql, **params):
@@ -62,42 +51,6 @@ def create_table(table, sql, **params):
         _create_table(table.lower(), con, sql, **params)
 
 
-def create_raw_tables():
-    engine = get_engine()
-    with engine.begin() as connection:
-        for filetype in ["activity", "organisation"]:
-            table_name = f"_raw_{filetype}"
-            logger.debug(f"Creating table: {table_name}")
-            connection.execute(
-                text(
-                    f"""
-                    DROP TABLE IF EXISTS {table_name};
-                    CREATE TABLE {table_name}(
-                        id SERIAL, prefix TEXT, dataset TEXT, filename TEXT, error TEXT, version TEXT, object JSONB
-                    );
-                    """
-                )
-            )
-
-
-@functools.lru_cache
-def get_xml_schema(filetype: str) -> xmlschema.XMLSchema10:
-    if filetype == "activity":
-        return xmlschema.XMLSchema(
-            str(
-                pathlib.Path()
-                / "__iatikitcache__/standard/schemas/203/iati-activities-schema.xsd"
-            )
-        )
-    else:
-        return xmlschema.XMLSchema(
-            str(
-                pathlib.Path()
-                / "__iatikitcache__/standard/schemas/203/iati-organisations-schema.xsd"
-            )
-        )
-
-
 def flatten_schema_docs(cur, path=""):
     for field, value in cur.items():
         info = value.get("info")
@@ -107,21 +60,6 @@ def flatten_schema_docs(cur, path=""):
             yield f"{path}{field}_{attribute}", attribute_docs
 
         yield from flatten_schema_docs(value.get("properties", {}), f"{path}{field}_")
-
-
-class IATISchemaWalker(sort_iati.IATISchemaWalker):
-    def __init__(self):
-        self.tree = etree.parse(
-            str(
-                pathlib.Path()
-                / "__iatikitcache__/standard/schemas/203/iati-activities-schema.xsd"
-            )
-        )
-        self.tree2 = etree.parse(
-            str(
-                pathlib.Path() / "__iatikitcache__/standard/schemas/203/iati-common.xsd"
-            )
-        )
 
 
 def get_schema_docs():
@@ -134,35 +72,6 @@ def get_schema_docs():
         schema_docs_lookup[field] = [num + 10, doc]
 
     return schema_docs_lookup
-
-
-def get_sorted_schema_dict():
-    schema_dict = IATISchemaWalker().create_schema_dict("iati-activity")
-    return schema_dict
-
-
-def sort_iati_element(
-    element: etree._Element, schema_subdict: OrderedDict[str, OrderedDict]
-) -> None:
-    """
-    Sort the given elements children according to the order of schema_subdict.
-    """
-
-    def sort(x: etree._Element) -> int:
-        try:
-            return list(schema_subdict.keys()).index(x.tag)
-        except ValueError:
-            # make sure non schema elements go to the end
-            return 9999
-
-    children = list(element)
-    for child in children:
-        element.remove(child)
-    for child in sorted(children, key=sort):
-        element.append(child)
-        subdict = schema_subdict.get(child.tag)
-        if subdict:
-            sort_iati_element(child, subdict)
 
 
 ALL_CODELIST_LOOKUP = {}
@@ -206,42 +115,6 @@ def get_codelists_lookup():
                 ALL_CODELIST_LOOKUP[(path, codelist_value)] = value_name
 
 
-def parse_dataset(
-    dataset: iatikit.Dataset,
-) -> Iterator[tuple[dict[str, Any], list[xmlschema.XMLSchemaValidationError]]]:
-    try:
-        dataset_etree = dataset.etree.getroot()
-    except Exception:
-        logger.debug(f"Error parsing XML for dataset '{dataset.name}'")
-        return
-
-    version = dataset_etree.get("version", "1.01")
-    if version.startswith("1"):
-        logger.debug(f"Transforming v1 {dataset.filetype} file")
-        dataset_etree = VERSION_1_TRANSFORMS[dataset.filetype](dataset_etree).getroot()
-
-    parent_element_name = (
-        "iati-organisations"
-        if dataset.filetype == "organisation"
-        else "iati-activities"
-    )
-    child_element_name = f"iati-{dataset.filetype}"
-    for child_element in dataset_etree.findall(child_element_name):
-        sort_iati_element(child_element, get_sorted_schema_dict())
-        parent_element = etree.Element(parent_element_name, version=version)
-        parent_element.append(child_element)
-
-        xmlschema_to_dict_result: tuple[dict[str, Any], list[Any]] = xmlschema.to_dict(
-            parent_element,  # type: ignore
-            schema=get_xml_schema(dataset.filetype),
-            validation="lax",
-            decimal_type=float,
-        )
-        parent_dict, error = xmlschema_to_dict_result
-        child_dict = parent_dict.get(child_element_name, [{}])[0]
-        yield child_dict, error
-
-
 def csv_file_to_db(csv_fd):
     engine = get_engine()
     with engine.begin() as connection:
@@ -249,100 +122,6 @@ def csv_file_to_db(csv_fd):
         copy_sql = "COPY _all_activities(prefix, dataset, filename, error, version, activity) FROM STDIN WITH CSV"
         cur = dbapi_conn.cursor()
         cur.copy_expert(copy_sql, csv_fd)
-
-
-def load_dataset(dataset: iatikit.Dataset) -> None:
-    if not dataset.data_path:
-        logger.warn(f"Dataset '{dataset}' not found")
-        return
-
-    path = pathlib.Path(dataset.data_path)
-    prefix, filename = path.parts[-2:]
-
-    engine = get_engine()
-
-    with engine.begin() as connection:
-        connection.execute(
-            insert(
-                table(
-                    f"_raw_{dataset.filetype}",
-                    column("prefix"),
-                    column("dataset"),
-                    column("filename"),
-                    column("error"),
-                    column("version"),
-                    column("object"),
-                )
-            ).values(
-                [
-                    {
-                        "prefix": prefix,
-                        "dataset": dataset.name,
-                        "filename": filename,
-                        "error": "\n".join(
-                            [f"{error.reason} at {error.path}" for error in errors]
-                        ),
-                        "version": dataset.version,
-                        "object": json.dumps(object),
-                    }
-                    for object, errors in parse_dataset(dataset)
-                ]
-            )
-        )
-
-    engine.dispose()
-
-
-def create_database_schema():
-    if schema:
-        engine = get_engine()
-        with engine.begin() as connection:
-            connection.execute(
-                text(
-                    f"""
-                    DROP schema IF EXISTS {schema} CASCADE;
-                    CREATE schema {schema};
-                    """
-                )
-            )
-
-
-def load(processes: int, sample: Optional[int] = None) -> None:
-    create_database_schema()
-    create_raw_tables()
-
-    logger.info(
-        f"Loading {len(list(islice(iatikit.data().datasets, sample)))} datasets into database"
-    )
-    datasets_sample = islice(iatikit.data().datasets, sample)
-
-    with concurrent.futures.ProcessPoolExecutor(max_workers=processes) as executor:
-        future_to_dataset = {
-            executor.submit(load_dataset, dataset): dataset
-            for dataset in datasets_sample
-        }
-        for future in concurrent.futures.as_completed(future_to_dataset):
-            # We have to get the result (even though we don't use it) in order to get the exceptions
-            try:
-                future.result()
-            except Exception as e:
-                dataset = future_to_dataset[future]
-                logger.error(f"Dataset '{dataset.name}' caused error {e}")
-
-    engine = get_engine()
-    with engine.begin() as connection:
-        activity_result = connection.execute(
-            text("SELECT COUNT(*) AS count FROM _raw_activity")
-        ).first()
-        logger.info(
-            f"Loaded {activity_result.count if activity_result else 0} activities into database"
-        )
-        organisation_result = connection.execute(
-            text("SELECT COUNT(*) AS count FROM _raw_organisation")
-        ).first()
-        logger.info(
-            f"Loaded {organisation_result.count if organisation_result else 0} organisations into database"
-        )
 
 
 def process_registry() -> None:
